+++ react-components/src/components/ModalMenu.jsx
@@ -100,7 +100,7 @@ export function Menu(props) {
         <hr className="m-v-xxs" />
         <li>
           <a
-            href="https://www.great.gov.uk/contact/feedback/"
+            href="/contact-us/help/"
             target="_blank"
             rel="noopener noreferrer"
             className="link"
@@ -110,11 +110,7 @@ export function Menu(props) {
           </a>
         </li>
         <li>
-          <button 
-            type="button" 
-            className="link" 
-            onClick={logout}
-          >
+          <button type="button" className="link" onClick={logout}>
             <i className="fa fa-arrow-right" />
             <span>Sign out</span>
           </button>
@@ -125,7 +121,7 @@ export function Menu(props) {
       <ul className="menu-items">
         <li>
           <a
-            href="https://www.great.gov.uk/contact/feedback/"
+            href="/contact-us/help/"
             target="_blank"
             rel="noopener noreferrer"
             className="link"